Refactor inplace() and range_indexed() decorators to own file
